réseau marche + calcule horaires marche

+++ centralWorkspace/pointeuse/src/pointeuse/EmulatorController.java
@@ -23,59 +23,70 @@ public class EmulatorController {
 		thread.start();
 	}
 
+	public TimeManagerController getTimeManager() {
+		return time;
+	}
+
 	public void sendId(int id) {
-		String message = createMessage(id);
-		System.out.println(message);
 
-		Socket socket;
-		OutputStream outputStream;
-		ObjectOutputStream objectOutputStream;
+		if (!time.getTodayWeekDay().equals("Sunday") && !time.getTodayWeekDay().equals("Saturday")) {
+			System.out.println(time.getTodayWeekDay());
+			String message = createMessage(id);
+			System.out.println(message);
 
-		// Initialisation du socket d'envoie
-		try {
-			socket = new Socket("127.0.0.1", 6890);
-			// Flux de donnees:
-			outputStream = socket.getOutputStream();
-			// Envoie des donnes via objectOutputStream:
-			objectOutputStream = new ObjectOutputStream(outputStream);
+			Socket socket;
+			OutputStream outputStream;
+			ObjectOutputStream objectOutputStream;
 
-			ArrayList<String> infos = readData();
+			// Initialisation du socket d'envoie
+			try {
+				socket = new Socket("127.0.0.1", 7720);
+				// Flux de donnees:
+				outputStream = socket.getOutputStream();
+				// Envoie des donnes via objectOutputStream:
+				objectOutputStream = new ObjectOutputStream(outputStream);
 
-			if(infos != null) {
-				for(String info : infos) {
-					objectOutputStream.writeObject(info);
+				ArrayList<String> infos = readData();
+
+				if (infos != null) {
+					for (String info : infos) {
+						objectOutputStream.writeObject(info);
+					}
+					deleteFile();
 				}
-				deleteFile();
+
+				objectOutputStream.writeObject(message);
+
+				objectOutputStream.close();
+				outputStream.close();
+				socket.close();
+			} catch (IOException e) {
+
+				saveData(message);
 			}
 
-			objectOutputStream.writeObject(message);
-
-			objectOutputStream.close();
-			outputStream.close();
-			socket.close();
-		} catch (IOException e) {
-			saveData(message);
+			System.out.println(readData());
+		} else {
+			System.out.println("ENVOYER UN MESSAGE DERREUR ICI");
 		}
-
-		System.out.println(readData());
 	}
 
 	public void deleteFile() {
 		try {
 			File file = new File("save.txt");
-			if(file.exists())
+			if (file.exists())
 				Files.delete(file.toPath());
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
 	}
 
-	public ArrayList<String> readData(){
+	public ArrayList<String> readData() {
 		ArrayList<String> infos = new ArrayList<String>();
 
 		try {
 			File file = new File("save.txt");
-			if(file.exists()) {
+			if (file.exists()) {
 				Scanner reader = new Scanner(file);
 				while (reader.hasNextLine()) {
 					infos.add(reader.nextLine());
@@ -92,9 +103,9 @@ public class EmulatorController {
 	public void saveData(String info) {
 		try {
 			File file = new File("save.txt");
-			if(!file.exists())
+			if (!file.exists())
 				Files.createFile(file.toPath());
-			Files.write(file.toPath(), (info +'\n').getBytes(), StandardOpenOption.APPEND);
+			Files.write(file.toPath(), (info + '\n').getBytes(), StandardOpenOption.APPEND);
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
